% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbchly.R
\name{nbchly}
\alias{nbchly}
\title{This function is allows you to calculate the number of species per chorology in a floristic data table}
\usage{
nbchly(x)
}
\arguments{
\item{x}{data matrix of floristic inventory, the name of each colum is as follow : 'genera', 'epit.spec', 'specie', 'family', 'life.form', 'chorology'.}
}
\value{
to a data.frame of number of species in each chorology origin
}
\description{
This function is allows you to calculate the number of species per chorology in a floristic data table
}
\author{
Evans Ehouman
}
