% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbgr.R
\name{nbgr}
\alias{nbgr}
\title{This function is allows you to calculate the number of species per genra in a floristic data table}
\usage{
nbgr(x)
}
\arguments{
\item{x}{data matrix of floristic inventory, the name of each colum is as follow : 'genera', 'epit.spec', 'specie', 'family', 'life.form', 'chorology'.}
}
\value{
to a data.frame of number of species in each genera
}
\description{
This function is allows you to calculate the number of species per genra in a floristic data table
}
\author{
Evans Ehouman
}
\keyword{family}
