% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblfm.R
\name{nblfm}
\alias{nblfm}
\title{This function is allows you to calculate the number of species per life form in a floristic data table}
\usage{
nblfm(x)
}
\arguments{
\item{x}{data matrix of floristic inventory, the name of each colum is as follow : 'genera', 'epit.spec', 'specie', 'family', 'life.form', 'chorology'.}
}
\value{
to a data.frame of number of species for each life form
}
\description{
This function is allows you to calculate the number of species per life form in a floristic data table
}
\author{
Evans Ehouman
}
\keyword{family}
