% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaRich.R
\docType{package}
\name{TaxaRich}
\alias{TaxaRich}
\alias{TaxaRich-package}
\title{TaxaRich Package
TaxaRich: Calculate taxa richness}
\description{
TaxaRich Package
TaxaRich: Calculate taxa richness
}
\author{
Ehouman Evans \href{mailto:ehoumanevans@gmail.com}{ehoumanevans@gmail.com}
}
